moved operation structs out.
