+++ xml-mut-xot/src/lib.rs
@@ -2,6 +2,7 @@ mod errors;
 mod fitable;
 mod macros;
 mod node_ext;
+mod operation;
 mod valuable;
 
 pub mod prelude {
@@ -9,9 +10,6 @@ pub mod prelude {
     pub use super::fitable::*;
     pub use super::macros::*;
     pub use super::node_ext::*;
+    pub use super::operation::*;
     pub use super::valuable::*;
 }
-
-// TODO: instead of direct replacer.
-// define operations to be performed.
-//
